requirements.txt
pytorch
transformers
datasets
scikit-learn
FastAPI
Uvicorn